refactor: extract UI update helpers from app.tsx
Move flushAssistantUpdates, updateAssistant, and updateTool into
src/app/ui-updates.ts as a createUiUpdaters factory.

App now delegates batching text/reasoning deltas and tool status patches
to the extracted module. Zero behavioral change.

Commit-Queue-Session: cq_20260508160558_dad111b7260fef0bcd71f2b9
Coding-Agent: pi
Coding-Agent-Session: pi-019e06b6-830e-72ca-8378-e2ba2d3209c0
